Update similarity.py
Changed input of target hash to be a single value instead of multiple values as array. This was done to avoid further complication of acceptable data type.
Now, The value_src and value_query is expected to be an integer or hexadecimal or array of binary or array of ints